CNNSeq2exp

Documentation for CNNSeq2exp.

CNNSeq2exp.motifs

CNNSeq2exp.sequence_data

CNNSeq2exp.get_num_seq_segments_and_heads

CNNSeq2exp.make_fasta_file

CNNSeq2exp.make_new_permuted_indices!

CNNSeq2exp.num_mega_batches_train

CNNSeq2exp.one_hot_to_seq

CNNSeq2exp.render_motifs

CNNSeq2exp.take_pretrained_and_render

CNNSeq2exp.train_and_render

CNNSeq2exp.train_model

CNNSeq2exp.motifs — Type

pfms: A vector where... {1st, 2nd,...} component : key is the {1st, 2nd,...} entry in keyst value is the {1st, 2nd,...} entry in valuest stores the pfms for the training set similarly for pfmstrainshuffled, pfmstest, pfmstest_shuffled

CNNSeq2exp.sequence_data — Type

fid: hdf5 file seqlen: length of the sequence (assume for now that all sequences have the same length) numseqsegments: number of sequence segments (number of input sequences) numheads: number of heads (number of differny y values) numseqs: number of sequences in the dataset permuteindices_train: indices for permuting the sequence (for training)

CNNSeq2exp.get_num_seq_segments_and_heads — Method

Get the number of sequence segments and heads from the hdf5 file. fid: hdf5 file, obtain via, e.g., h5open(<filepath>, "r")

CNNSeq2exp.make_fasta_file — Method

Take in a vector of (vector of two strings) 1st string: header 2nd string: sequence and then make a fasta file

CNNSeq2exp.make_new_permuted_indices! — Method

Make new permuted indices for the training data. sd: sequencedata loadhow_many: how many data entries to load each time

CNNSeq2exp.num_mega_batches_train — Method

Get the number of mega batches for training. nummegabatches: number of mega batches sd: sequence_data

CNNSeq2exp.one_hot_to_seq — Method

Take an one-hot encoded sequence and convert it to ACGT string

CNNSeq2exp.render_motifs — Method

Render the motifs

CNNSeq2exp.take_pretrained_and_render — Method

fp: file path for the formatted hdf5 file; that contains the sequence and label information, hploadpath: path to the hyperparameters file; this file should be in the jld2 format modelloadpath: path to the model file; this file should be in the jld2 format save_where: a folder path to save the results; note that the folder will be created if it does not exist. UPTO: maximum number of filters in the coalition to be considered.

CNNSeq2exp.train_and_render — Method

fp: file path for the formatted hdf5 file; that contains the sequence and label information, save_where: a folder path to save the results; note that the folder will be created if it does not exist. UPTO: maximum number of filters in the coalition to be considered.

CNNSeq2exp.train_model — Method

Inputs: fp: file path hp: hyperparameters if nothing, then random hyperparameters are generated lastembeddinglenmax: max length of the last embedding nepochs: number of epochs to train the model Outputs: hp: hyperparameters m: trained model